+++ CHANGELOG.md
@@ -9,12 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 These are changes that will probably be included in the next release.
 
 ### Added
+ * Add DisableLogger and EnableLogger to disable/enable specific named loggers.
  
+
 ### Fixed
 
 ### Changed
  
 ### Removed
+ * Deprecated Null logger and it will be removed in next version, use logr.Discard instead
+ * Deprecated FromContext and NewContext and they will be removed in next version, use logr.FromContext and logr.NewContext 
+
 
 ## [v0.1.4] - 2020-10-11